console/backend/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "envbasis-backend"
version = "0.1.0"
description = "Backend API and data layer for Envbasis."
readme = "plan.md"
requires-python = ">=3.9,<4.0"
dependencies = [
  "cryptography>=42.0,<44.0",
  "fastapi>=0.110,<1.0",
  "uvicorn[standard]>=0.27,<1.0",
  "sqlalchemy>=2.0,<3.0",
  "alembic>=1.13,<2.0",
  "psycopg[binary]>=3.1,<4.0",
  "PyJWT>=2.8,<3.0",
  "pydantic-settings>=2.2,<3.0",
  "email-validator>=2.1,<3.0",
  "python-dotenv>=1.0,<2.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0,<9.0",
]

[tool.setuptools.packages.find]
include = ["app*"]
